// sourcecode/BTrak.Packages/snova-audits-module/projects/launcher.dev/src/app/app.module/models/dragedWidget.ts
export class DragedWidget {
    name: string;
    isCustomWidget: boolean;
    customWidgetId: string;
    widgetId: string;
    visualizationType: string;
    xCoOrdinate: string;
    yCoOrdinate: string;
    customAppVisualizationId: string;
    isHtml: boolean;
    isProc: boolean;
    procName: string;
    isProcess: boolean;
    isEntryApp: boolean;
    isEditable: boolean;
}